factor out asset reconciliation graph traversal into util
branch-name: reconciliation-sensor-refactor
